refactor: apply feedback of ticker-detail page feature implementation - divide LiveTradingChart component, change variables and check EOL
