+++ src/widgets/ticker-detail/lib/constants.ts
